requirements.txt
fastapi
uvicorn
camelot-py[cv]
pandas
ghostscript
opencv-python-headless
